Updated openstack/openstack Project: openstack/tempest  509de5a994e5d75c5fbee62b3dd1efdf392cc36b
Adds missing server tags APIs to servers client.

Currently, there is no server tags API testing implemented in
Tempest [1]. This API was introduced to Nova in 2.26 [2].

This patch adds the server tags APIs to the servers client,
as well as corresponding unit test and schema changes.

Mocking was added for the unit tests, because these tests
explicitly require that the COMPUTE_MICROVERSION header
in the base_compute_client be '2.26', or else the schema
that is returned by get_schema in base_compute_client
will return the v2.1 schema, resulting in validation errors.

[1] https://github.com/openstack/tempest/tree/master/tempest/lib/services/compute
[2] https://developer.openstack.org/api-ref/compute/

Change-Id: I06b9c6f42e066863310b4fe7ee7030c1dfb89467
Implements: blueprint missing-server-tags-api-test