Updated openstack/openstack Project: openstack/keystone  75f12a35d77b3d2bed184b70cb7ff9a235fe642a
Add instruction to restart apache

It can be easy, even for seasoned operators, to either forget to restart the
web server or to cautiously avoid restarting the webserver unless instructed
to do so. Since there are notes all over the rest of the documentation to
restart the web server after making configuration changes, add a similar note
to the WebSSO section for consistency.

Change-Id: I1d171a2f0ccc2b70fe3e2e48fe2f2303c515b9a6