Updated openstack/openstack Project: openstack/nova-specs  ad4c1c5fe77818e2c788d58bbacaca91c138a508
Update requirements to fix pbr 2.0.0 fail

nova-specs doesn't sync requirements updates from the
global requirements repo and doesn't have the correct
requiremnts ranges for hacking which is pulling in
pbr 2.0.0 which breaks with warnerrors=True (since that
was removed in 2.0.0).

This syncs up the requirements with the latest from
the requirements repo and updates the hacking versions.

Change-Id: I8db1038693c5047ebffd0065055efc5ea211b106
Related-Bug: #1668848