Updated openstack/openstack Project: openstack/os-collect-config  ec8b2576daddaa5d584256de42beb72ecdd690d5
Update to a newer hacking library

The older hacking library has a cap on pbr <2.0, with the recent 2.0.0
release of PBR it's causing failures in the pep8 job.

hacking isn't kept in sync via the typical proposal-bot updates.

Do it manually to clear the gate issue.

pep8 runtests: commands[0] | flake8
./os_collect_config/merger.py:45:22: W292 no newline at end of file
    return final_list

Change-Id: If863d2e190a4e1701ebf1fcb99a783075cda2556
Related-Bug: #1668848