Updated openstack/openstack Project: openstack-dev/devstack  93715a93df93dc81d2765b3f794f7cdcf51d67b9
clean up virt drivers that we do image logic for

openvz is not in the nova tree, and is referencing a crazy old image,
we're going to assume that if anyone is using this they can build a
devstack plugin.

drop doing anything by default because this actually requires that we
special case things like ironic in tree to *not* do anything by
default.

Change-Id: I9d33b98263c3d52a95b9983e90eb0b341fa1d363