Updated openstack/openstack Project: openstack/neutron  55d3c9a79db01979fd3f40282e2e519f1b18d59e
Scan for mac through all devices

The first device the agent happens to pick could be something
without a mac address like a 6in6 interface. This was causing
the agent to fail to start if it was unlucky enough to pick
that address.

This patch just adjusts the logic to keep iterating through
the list until we find a mac.

Closes-Bug: #1669087
Change-Id: I6e934a2dff8fd441e1c70c9a5857fd6150835e2a