Updated openstack/openstack Project: openstack/horizon  55c77bdb59be390ae30ea771a6197377daf88ca4
Merge "missing" quotas with nova quotas

There was an extra variable for the so called "missing"
quotas since 2013. This patch merges them with the
normal nova quotas, cleans up the tests, removes any
usage of that special var and as a side effect, adds
the missing keys to the project quota update page.

Change-Id: Icaac1ebb1749e13fdfc307e89dcc6bbd64bef922
Closes-Bug: 1655393