Updated openstack/openstack Project: openstack/neutron-vpnaas  1609ecfd2fb5303644756e4510d4204574f4eac6
Update a py35 environment to tox and classifier

Infra has been running python35 jobs, replaced the python34 jobs.
This is due to the change from Ubuntu Trusty to Xenial, where only
python3.5 is available.

So we update a py35 environment with the same definitions for
skipping not-yet-working checks as we have in place for py34 already.

Change-Id: I3e7fc720ca6bd9df80620d609e9a9603a644673a