Updated openstack/openstack Project: openstack-infra/project-config  31c8b20d3ce21a8e3412131acb64c3262f28875f
layout/tripleo: fix critical error in the job name regex

I made a mistake in this patch:
I3df771d98adffde272df94661fcc31aec4661fde

Where the regex became too general for the end of the job name.
This patch aims to fix it, and make sure this layout only works for
multinode and multinode-nv jobs, and not multinode-puppet and
multinode-puppet-nv jobs.

Change-Id: I049d8b11dd3c52d621fdfa02f8f768be840c0174