Updated openstack/openstack Project: openstack/cinder  892d5989cf149031cec5563d408ddf28007d88ff
Enforce _usage_from_snapshot to fetch info from deleted volume

The fix introduced for LP: #1631561 enforces the availability_zone to be
a empty string if the volume cannot be loaded because it has been marked
as deleted. This change enforces the load of the deleted volume in order
to load the volume's availability_zone.

Change-Id: I926f90d49fa0b352975232a68be5fbb2f8304d3b
Closes-Bug: #1666686