Updated openstack/openstack Project: openstack/horizon  79128d5ba9df8ab847c164ab29fd8f91e7745175
Close modals with ESC key

Currently, when the user opens a modal window and clicks on an element
that is NOT a form input and tries to close the modal window by pressing
the ESC key, the modal window will not close. This patch allows the user
to close modals by pressing the ESC key no matter what element on the
page has the focus.

The fix is taken from https://getbootstrap.com/javascript/#live-demo

Co-Authored-By: Rob Cresswell <robert.cresswell@outlook.com>

Change-Id: I4d19fb400dee8132d9d490507c06989c07ce42ac
Closes-Bug: #1561193