Updated openstack/openstack Project: openstack/horizon  fc8c13231ab8525e42911e36bfa41d96d05c978d
fix workflow exception on host aggregate create form

Under the circumstance that there are existing aggregates, this bug
happens when create button is clicked without the required "name" field
filled. Error toast shows instead of the expected error message on the
create form.
When the form posts without name filed filled, the name attribute in the
clean data is None instead of an empty string, which causes
AttributeError on the following string operation.

1. patch an empty string as the fallback value for the 'name' variable
2. corresponding unit test is attached

Change-Id: I0bdf469d87aa06ab58b73e6ce41bdf63bb36a538
Closes-Bug: #1617140