Updated openstack/openstack Project: openstack/oslo.utils  64468e5c3cfeeeec83172c130b5f8f01e1e824d5
Add missing documentation for secretutils

The current doc page for secretutils is blank [1], most likely because
of the way constant_time_compare is in the exception block.

This patch manually tells sphinx to doc the constant_time_compare
function. Also adds the oslo.utils version when the module was added.

[1]: https://docs.openstack.org/developer/oslo.utils/api/secretutils.html

Change-Id: Ic1fb15176c93f126826f568149cabf4b8e5450d8