Updated openstack/openstack Project: openstack/nova-specs  eeec1cb806d21102bb88893ab319d122e5c2429f
Libvirt driver emulator threads placement policy

The Nova schedular determines CPU resource utilization and instance
CPU placement based on the number of vCPUs in the flavour. A number of
hypervisors have work that is performed in the host OS on behalf of a
guest instance, which does not take place in association with a
vCPU. This is currently unaccounted for in Nova scheduling and cannot
have any placement policy controls applied.

There are some updates to the spec based on changes to the design
during Ocata, and the History section is added.

Previously-approved: ocata
Change-Id: I4d65e4ee613236f80205d4d61edf3fcf8bd1fe4b
Blueprint: libvirt-emulator-threads-policy