Updated openstack/openstack Project: openstack/cinder  1ca1c47cd0f44cf033c46e02fc2c6ca17c42f919
Remove v1 API reference

The v1 API has been deprecated for a long time, and we would like to
eventually be able to completely remove it. All new development should
be using v2 or v3, ideally the latter.

This removes the v1 API info from our api-ref in the hopes of further
discouraging API consumers from using the v1 API.

Change-Id: Ib14749fb6bae88e89af28493156b478c2099fd17