Updated openstack/openstack Project: openstack/cinder  65b07dfa4f01077cb81b35322c20739cc83287c7
api-ref: v2 update volume metadata description is wrong

Modify metadata description regarding update volume. The reasons are as follows.
In line 348, current description of metadata is talking about a snapshot rather
than a volume, and this metadata should be optional as part of a request body.
In line 376, the description of metadata is talking about a snapshot rather
than a volume, and this metadata should be required as part of a response.

Change-Id: I9647ce149760121f3140fe820f350aa2d8589ae0
Closes-Bug: #1668301