Updated openstack/openstack Project: openstack/nova-specs  af96769eb9e4603e835c0911926fe795e3342a22
CellsV2 - Move quota tables to the API db

This was worked on in Ocata but didn't land any changes, so re-proposed
for Pike. The original spec included use of the quota_usages and
reservations tables and that has since changed, so the spec has been
updated to reflect that.

Related to blueprint cells-quota-api-db

Change-Id: I4ebc01aead5caf75126085c842be26a870ec9c84