Updated openstack/openstack Project: openstack/nova  5cf6bbf374a8b877d2e158aa8802b31d14ceb121
Add warning on setting secure_proxy_ssl_header

secure_proxy_ssl_header is used when the API is behind an proxy that
terminate the HTTPS connection.

This commit is just adding a warning on the config description in order
to make sure that the user understand what he is doing when setting
this.

Change-Id: I5c2cb56109508fb7215a3a42a0ab1c831083182f
Signed-off-by: Arnaud Morin <arnaud.morin@corp.ovh.com>