Updated openstack/openstack Project: openstack/neutron  b34f705536864ea0b7126cc027340d9ff6e5d1ac
Deprecate nova_metadata_ip in favor of nova_metadata_host option

nova_metadata_ip option name is a bit confusing as it accepts an ip but
also a dns name (which is required when https protocol is used).

This change deprecates nova_metadata_ip option in favor of
nova_metadata_host option and updates option help to highlight that we
can use an ip or a dns name.

DocImpact
Closes-Bug: #1585699
Change-Id: Ia6c2471c7b4f3e924941222133edbb90151757a5