Updated openstack/openstack Project: openstack/trove  3b64992806890e3c1f034c2e6e0d6176205edf08
Stabilize gate: Fix CONTROLLER_IP is .1 issue

There is still a bug in determining the host IP in a gate run
if it ends with .1 - this is due to the fact that the variable
has spaces in it so the -z test for a missing IP does not return
true.

This has been changed to ignore spaces.

Change-Id: I70701f1d3985b0347fd2aa796e712ead73ea9635
Closes-Bug: 1668674