Updated openstack/openstack Project: openstack/tripleo-heat-templates  47fcabee464b1e46490087316778b3cff0a55b5b
Put docker puppet config in puppet_config dict

This approach removes the need for the yaql zip to build the
docker-puppet data by building the data in a puppet_config dict.

This allows a future change to make docker-puppet.py only accept dict
data.

Currently the step_config is left where it is and referenced inside
puppet_config, but feedback is welcome whether this is necessary or
desirable.

Change-Id: I4a4d7a6fd2735cb841174af305dbb62e0b3d3e8c