Updated openstack/openstack Project: openstack/keystone-specs  3f1d3057e992660819891c5fd2bb1fa205b790f6
Remove centralized policy delivery spec from backlog

This was something we discussed during a recap of the Pike PTG:

  http://eavesdrop.openstack.org/meetings/keystone/2017/keystone.2017-02-28-18.00.log.html#l-231

There are a lot of policy discussions in flight the conflict with the
ideas specified in this specification. Which ever approach we decide
on as a community will likely need a new proposal.

Removing this from the backlog so that we don't mislead new
developers or people who frequently reference our backlog to gauge
project direction.

Change-Id: Ida5387ceef2823a2d54d94313fca783ba518e54d