Updated openstack/openstack Project: openstack-infra/devstack-gate  0e107ea76a386215349d3497596bc04eacaa567f
Run test-matrix.py under python3 if present

This checks for the presence of python3 and if present installs the
python3 deps for test-matrix.py and then runs test-matrix.py under
python3. This is useful because as OpenStack grows more python3
compatibility we remove a python2 dep in the test env allowing people to
potentially have python3 only test envs.

Change-Id: I9253f3c5875e8c06f5dcb90b7a6c1c95c05d89cb