Updated openstack/openstack Project: openstack/manila  54b030a3f2a93bc2d85bb9d0f8856f0938265c87
[Share Groups] Squash SGS member and SS instances DB tables

Because the former was implemented just because the latter
was absent yet at that moment. Main reason is that they serve
same goals and should be consistent to each other. And it is
much easier to support one single table instead of two separate
tables syncing them each time we update one of them.

Change-Id: I81ba26437554d02a79e4b536d781617f46ce2f07
Partially-Implements bp manila-share-groups