Updated openstack/openstack Project: openstack/os-apply-config  dcc453a946a98532e5f0b68ebcc52f96a56fb7cb
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove os-apply-config.

Change-Id: I792427363cda300650d2ce1c0fdab9973be7b348