Updated openstack/openstack Project: openstack/cinder  c356c92ad8b0e5050b661506bc8b769398f6ddb9
Disallow modification of in use Volume Types

Modifying a Volume Type doesn't actually update any
volumes that are currently using said Volume Type. Given
that this is an admin only operation it was never really
too much of a concern.

There has been some issues reported however where an admin
modified the type while volumes were using it, then ran into
problems becuase the volumes didn't have the expected settings
described by the type.

This change adds a check before updating/deleting extra-specs of
a volume type.  If there are any volumes currently assigned to
the type being modified/deleted the operation will fail with an
InvalidRequest immediately when attempting the call.

To maintain backward compatability incase someobdy is using this
for something and they are really really sure they want to
continue doing so, we also add a config option to allow the
old behavior but default to NOT allowing it:
  'allow_inuse_volume_type_modification=False'

APIImpact

Change-Id: Iaea721e13a3903cae60cc3fb3acfad03bd173a6b
Closes-Bug: #1667071