Updated openstack/openstack Project: openstack-infra/project-config  3c5895d3dde728ccc3656148ce36a4fe0875493a
[magnum] Refactor builder and job templates

* Add new job templates that allows to explicitly use specific
  projects.
* Copy the current builder to a new builder to be used by the new
  job templates.
* Add multinode k8s job and job template.

This patch generates a lot of job and job-template duplication to
ensure that the new model of jobs does what it's suppose to do.
The new jobs are added as non-voting for testing. Provided that
the results is satisfying, the old job templates will be removed
and the new jobs will become voting in a follow-up patch.

Change-Id: I56b97a244973ca4d94ff3cb92debe99afef8c559