Updated openstack/openstack Project: openstack/neutron  97db8aab076281c7b53a5ccef6e5c7f0939b3903
Get rid of ML2 inheritance of delete_network

In an effort to simplify the codebase, this eliminates
all of the custom code in the ML2 delete_network
method and leverages registry callbacks for network
events.

The only purpose of the remaining inheritance of the
method is to leave the transaction guard on.

Partial-Bug: #1666424
Change-Id: Ib7d1c2081654dda82b5e015c71d20f378f3308f7