Updated openstack/openstack Project: openstack/nova  d3cb8bbc18e665249a1e284f31325856690bcca2
Default firewall_driver to nova.virt.firewall.NoopFirewallDriver

In Ocata we changed use_neutron=True as the default. When configuring
Nova to use Neutron, one is supposed to set the firewall_driver
config option to nova.virt.firewall.NoopFirewallDriver, but that was
not the default. This change makes it the default to match the default
for use_neutron.

Change-Id: If012acdd0ef2a66948ae45288cccb766262efb68