Updated openstack/openstack Project: openstack/neutron-lbaas  fb2d5fb1c89c004624a2cc3fa153c411b68ce517
tempest: Switch to local copy of tempset/scenario/manager.py

The manager.py file was mostly literally copied from
tempest as of today. [1]  It's slightly modified to
appease N521 hacking check.

tempest/scenario/manager.py is not a part of tempest stable
interface for plugins. [2]

This was suggested on openstack-dev as a preparation for
the coming refactoring in tempest. [3]

As tempest is branchless, this change needs to be backported
to all relevant stable branches.

[1] d184d619d7adc6ab60055884515055a7f3cf7770
[2] https://docs.openstack.org/developer/tempest/plugin.html#stable-tempest-apis-plugins-may-use
[3] http://lists.openstack.org/pipermail/openstack-dev/2017-February/112938.html

Change-Id: I76466d8365b24d55c299dafb36d53bca919475aa