Updated openstack/openstack Project: openstack-infra/tripleo-ci  11745dea4d1664c8a6685eeb060e85bb7047d0cd
Add new variable REVIEW_RELEASE and use it in delorean build

STABLE_RELEASE variable is mutated in upgrade jobs to point to
the release used for initial deployment. However, we need to keep
the release of the review to be used in delorean build phase, as
currently it can lead to use the wrong release when building the
package.

Change-Id: I3dc9ae79700773315e2577239839d35af7b85987