Updated openstack/openstack Project: openstack/tempest  0cfd2591e6f278adae6d13edb5882f44ea7adc6a
Deprecate resources_prefix and change rand_name()

Many projects are using rand_name() which is outside of tempest.lib.
The method specifies the prefix 'tempest', which is the default value
of tempest.conf, to tempest.lib's rand_name().
The prefix is useful to identify these resources are created by Tempest.
No projects change this configuration value, so this patch changes the
default value of rand_name() to 'tempest' and makes resources_prefix
option deprecated.

Change-Id: I624e2eb3954b6171fbbe00a4ed757bfac04b3eaf