Updated openstack/openstack Project: openstack/swift  7f953ce0b9649c8eb1c1e20348de5949c3a07e3c
Follow up the reconstructor test coverage

This is follow up for https://review.openstack.org/#/c/436522/

I'd like to use same assertion if it goes the same path.
Both Exception and Timeout will be in the exception log starts with
"Trying to GET". "Timeout" is an extra word appeared in the log.

And more, this adds assertions for the return value from the
get_response for error cases which should be as None.

Change-Id: Iba86b495a14c15fc6eca8bf8a7df7d110256b0af