Updated openstack/openstack Project: openstack/tripleo-puppet-elements  85f4604298e3618fafabb5af3152166f9a7ab6a2
Remove nova.conf truncate/refresh workaround

truncate-nova-config was added as a workaround for the following
bug.

  Duplicate sections in generated config
  https://bugs.launchpad.net/oslo.config/+bug/1568820

This workaround is no longer needed, as the root causes are now
fixed in both nova and oslo-config.

This workaround is also unused, as it was recently changed and
accidentally installed in the wrong location (nested one level too
deep in configure.d, and maxdepth is 1 in dib-run-parts).

For example, compare configure.d/ in stable/newton vs master
(circa ocata), and note the path differences.

  newton: http://paste.openstack.org/show/601802/
  ocata: http://paste.openstack.org/show/601806/

  newton: configure.d/40-truncate-nova-config
  ocata: configure.d/40-truncate-nova-config/40-truncate-nova-config

But it was a happy accident because this workaround stripped the
comments from /etc/nova/nova.conf, but now they're back as of ocata.

Related-Bug: #1568820
Change-Id: Iddae03f2d66cf534908dad884af55e2ae7d18d4b