Updated openstack/openstack Project: openstack/nova  5f16ce546f808b826fd044229f236d232141551b
libvirt: check if we can quiesce before volume-backed snapshot

The NFS job in the experimental queue is stacktracing when taking
a volume-backed snapshot because the image used in the job does
not have the qemu guest agent in it. This is not fatal because the
code just continues to snapshot without quiesce.

Since we have code that can tell if we can even attempt a quiesce
we should use that here before attempting the snapshot with quiesce,
else don't even attempt it so we don't stacktrace.

Note that this also adds a check such that if the image requires
quiesce as part of a snapshot and we can't honor that, we fail.

The stacktrace is intentionally left in here in case the snapshot
with quiesce fails because we checked if we could and only get to
that point if we should be able to quiesce but it failed, which
is an error that should be investigated.

Also, the unused quiesce parameter is removed from the
test_volume_snapshot_create_libgfapi test while in this code.

Change-Id: I685c592340dc8a3b7d32834a0114f2925dc4305c
Closes-Bug: #1629371