Updated openstack/openstack Project: openstack/nova  31a90c16aba9a23d0e3ccd51c44d8b633f406c39
Revert "fix usage of opportunistic test cases with enginefacade"

This reverts commit e59f95439b46a243ff4aee277397e17f28f79c45 removing
the unused EngineFacadeFixture class.

The patch that needed this (https://review.openstack.org/#/c/283364/)
was never merged to Nova. We only wanted this in Nova tree as a
temporary workaround for oslo.db limitations. Since then the related
bug in oslo.db was closed - folks are encouraged to use the fixtures
provided by the library:

https://docs.openstack.org/developer/oslo.db/api/oslo_db.sqlalchemy.test_fixtures.html

Related-Bug: #1548960

Change-Id: Ia9792604bb6ef65b9601fb212b2aeca16284b69f