Updated openstack/openstack Project: openstack/neutron-fwaas  9cbc94d7da5211d81f9922fd7ea0cd5e448faac0
Define in_namespace contextmanager

This change defines the contextmanager in_namespace[1]. It moves current
process in a network namespace (in __enter__) and moves it back in its
original network namespace (in _exit__) or kills current process if
__exit__ fails in order to ensure following commands will be executed
in the correct network namespace.

This change is an enabler to the Netlink solution to clean conntrack
entries.

[1] neutron_fwaas.privileged.utils

Partial-Bug: #1664294
Change-Id: I587257db8e1fce56a95f0db3dc4e0752751fdd81