Updated openstack/openstack Project: openstack/requirements  fccc463727983641852a1469d5d83f1024917b69
Remove vmware-nsxlib from the global requirements

At the moment we have too much churn in the library. This is
causing too many changes and breaking too many things.

We have decided that the best approach for the short term is
to work in a 1:1 relation ship with the current openstack version
and when the library is more stable return to the model where we
can consume a specific version.

Change-Id: I2d845ef41402efa6545fd9c4432a005ab7bb7b89