Updated openstack/openstack Project: openstack/tempest  3fb74fc1abdb67a71352b9cf61be75fcd0ab3a8e
Add dhcp_client argument to renew_lease()

renew_lease() is used in test_network_basic_ops only and it is not
necessary to have a config value of dhcp_client in RemoteClient.
This patch adds dhcp_client argument to renew_lease() with the same
default value as the config.

This is a step for making remote_client stable interface(tempest.lib)
and we need to remove CONF dependency from such module as possible
before defining stable interface.

NOTE: openstack/vmware-nsx only needs to follow this change.

Change-Id: I55f7cfb004169369ca63d1a691811eaed3155414