Updated openstack/openstack Project: openstack-infra/project-config  928bc193745fe0fc3b85b58c768cab41ff0591ec
Adding Nemesis project. A file intelligence API.

Nemesis is a file intelligence API which allows querying of files by
hash and if the file is known intellig ence about said file is returned
such as a Malware likelyness score, the file type and so on. If the file
is not known the user is requested to submit the file to the API where a
number of scalable background workers process the file looking at PE
headers, MIME type and so on as well as running it through a number of
pluggable anti-virus drivers.

The use case is not really for file system scanning, and obviously
heurirstics scanning is not available as there is no running of said
file (yet, maybe later dynamic analysis may be included but this is a
long term stretched goal). Instead the intended purpose is for
developers of websites, APIs or other applications to scan uploads /
submissions on the fly before publishing them, or you could create a
Swift middleware to scan all files before they are accepted and placed
onto object nodes.

I beleive this could be a useful value add / complimentary project to
the OpenStack eco system, and in anycase would use Keystone for auth and
so on.

As of yet there is no code, I will probably end up recycling some
previous code from a personal project I made last year, however the repo
itself should be created fresh for a new start. I am happy as part of
the first patch to ensure the test stuff as provided by cookiecutter is
included and no mass import of existing code will take place.

Change-Id: Ida135e77746d303d8aa894b534d34f3f15466113